+++ source/config.c
@@ -25,7 +25,8 @@
   CONFIG_VAR_INT(show_wanted_stars); \
   CONFIG_VAR_INT(disable_ped_spec); \
   CONFIG_VAR_INT(no_offscreen_despawn); \
-  CONFIG_VAR_INT(mobile_widgets);
+  CONFIG_VAR_INT(mobile_widgets); \
+  CONFIG_VAR_INT(fuzzy_seek);
 
 Config config;
 
@@ -59,6 +60,7 @@ int read_config(const char *file) {
   config.disable_ped_spec = 1;      // on by default (remove ped/character specular shine)
   config.no_offscreen_despawn = 1;  // on by default (cars/peds stay off-screen)
   config.mobile_widgets = 0;        // off by default = hide the mobile touch widgets
+  config.fuzzy_seek = 1;            // on by default (skip useless mpg123 data on seek)
 
   FILE *f = fopen(file, "r");
   if (f == NULL)